partialWaveFit: 'fitResult' does not inherit from 'TObject'
Remove the inheritance of 'rpwa::fitResult' from 'TObject'. There is
no difference concerning the storage in a 'TTree' without this. For
the direct storage of a 'fitResult' object in a 'TFile', which we
never used, one can still use the 'TFile::WriteObject' method. At the
same time remove the unneccessary 'public:' declaration before the
ClassDef.

Also replace the ROOT types (Double_t, Int_t, UInt_t, Bool_t) with
standard types, the type of the variable '_rank' was changed to be an
'unsigned int' (was signed 'int').