partialWaveFit: do not store names of production amplitudes
Replace the names of the production amplitudes by directly storing
the rank and the index to the wave the production amplitude
corresponds to. For the moment still use the names of the production
amplitudes by creating these from the available information. As the
ordering of the waves is equal to the ordering of the waves in the
integral matrices, the map '_normIntIndexMap' is equivalent to the
vector '_waveProdAmpIndices'. Remove the map.

Do not increase the ClassDef version for ROOT I/O as the previous
version was not yet merged to master.