partialWaveFit: store reflectivity of wave in fit result
Also store the reflectivity of each wave in the fit result. The
'waveInfoType' in 'fitResult' and 'pwaLikelihood' was adjusted
accordingly. This removes quite a number of calls to
'partialWaveFitHelper::getReflectivity' which needed to parse the
reflectivity from a string.

Do not increase the ClassDef version for ROOT I/O as the previous
version was not yet merged to master.