partialWaveFit: calculate weights in 'fitResult'
Add a new method 'spinDensityMatrixTimesAmplitudeMatrix' to the
'fitResult' class. This method calculates the product of the
spin-density matrix for a range of waves with an amplitude matrix. If
the amplitude matrix is, e.g., the normalisation integral, the method
can be used to calculate the intensity of a number of waves. By also
using an adapter class, which calculates the used matrix elements
from a vector of amplitudes, this new method can also be used to
calculate the weight for an event or for a particular subset of waves
in an event. The 'spinDensityMatrixTimesAmplitudeMatrix' method has
been crosschecked against the 'intensity' method and the
'weightEvents' executable.