pyInterface: enable usage of numpy from C++ side
Enable the usage of 'numpy' from the C++ side. A required definition
for the 'numpy' API version is added to the overall 'CMakeLists.txt'.
The include directory of 'numpy' is used while compiling the Python
bindings. The array part of 'numpy' is initialized in the main
routine exporting the bindings to Python so it can be used in the
'pyRootPwa.core' module.